Additional cleanup:
* Avoid repeated security exceptions by caching empty method lists
* Localize more method-gathering state and logic into
  MethodGatherer class
* Clean up some visibility and organization.
